+++ gamemodes/terrortown/gamemode/client/vgui/cl_sb_row.lua
@@ -31,7 +31,8 @@ local dev_tbl = {
 	["76561198058039701"] = true, -- saibotk
 	["76561198047819379"] = true, -- Mineotopia
 	["76561198052323988"] = true, -- LeBroomer
-	["76561198076404571"] = true -- Histalek
+	["76561198076404571"] = true, -- Histalek
+	["76561198296468397"] = true -- ZenBre4ker
 }
 
 local vip_tbl = {
@@ -52,7 +53,8 @@ local vip_tbl = {
 	["76561198114719750"] = true, -- Reispfannenfresser
 	["76561198082931319"] = true, -- Henk
 	["76561198049910438"] = true, -- Zzzaaaccc13
-	["76561198296468397"] = true -- ZenBre4ker
+	["76561198296468397"] = true, -- ZenBre4ker
+	["76561198041170748"] = true -- Pytho | Paul
 }
 
 local addondev_tbl = {