+++ gamemodes/terrortown/gamemode/server/sv_player_ext.lua
@@ -1441,6 +1441,11 @@ local function SetPlayerReady(_, ply)
 
 	entspawnscript.TransmitToPlayer(ply)
 
+	-- update playermodels on the client
+	if ply:IsSuperAdmin() then
+		playermodels.StreamToSelectedClients(ply)
+	end
+
 	---
 	-- @realm server
 	hook.Run("TTT2PlayerReady", ply)